backend/requirements.txt
Django>=5.1,<5.2
psycopg2-binary>=2.9
django-ninja>=1.3.0
channels[daphne]>=4.2.0
channels-redis>=4.2.0
django-allauth>=65.3.0
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
django-extensions>=3.2.3
celery>=5.5.0
redis>=5.2.0
gunicorn>=23.0.0
python-dotenv>=1.0.1
# Pillow skipped due to network

requests>=2.31.0
requests>=2.31.0
PyJWT[crypto]>=2.8.0
requests>=2.31.0
PyJWT[crypto]>=2.8.0
